backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "endee>=0.1.18",
    "fastapi>=0.135.1",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
